+++ src/index.ts
@@ -24,7 +24,8 @@ export type {
   MjolnirGestureEvent,
   MjolnirKeyEvent,
   MjolnirWheelEvent,
-  MjolnirPointerEvent
+  MjolnirPointerEvent,
+  WheelDevice
 } from './types';
 
 export type {