+++ test/hammerjs/index.ts
@@ -3,3 +3,4 @@
 // Copyright (c) vis.gl contributors
 
 import './double-click-drag.spec';
+import './trackpad.spec';