+++ test/inputs/index.ts
@@ -2,6 +2,8 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) vis.gl contributors
 
+import './contextmenu-input.spec';
 import './key-input.spec';
 import './move-input.spec';
+import './wheel-gesture-session.spec';
 import './wheel-input.spec';