+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+#### v3.1.0-beta.2 - July 21, 2026
+
+- Added opt-in trackpad gesture support to two-pointer `Pan` and `Pinch` recognizers.
+- Added wheel device classification through `MjolnirWheelEvent.device`.
+
 #### v3.0.1 - Jul 3, 2026
 
 - feat: add DoubleClickDrag recognizer (#13)